pyproject.toml
[project]
name = "passconvert"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pykeepass>=4.1.1.post1",
    "python-gnupg>=0.5.4",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pre-commit>=4.2.0",
    "ruff>=0.11.8",
]

[tool.mypy]
ignore_missing_imports = true
strict = true

[tool.ruff]
line-length = 100
target-version = "py311"
